+++ src/typeorm/entities/announcement/AnnouncementConfig.ts
@@ -1,7 +1,6 @@
-import { Column, CreateDateColumn, Entity, Index, PrimaryGeneratedColumn, UpdateDateColumn } from 'typeorm';
+import { Column, CreateDateColumn, Entity, PrimaryGeneratedColumn, UpdateDateColumn } from 'typeorm';
 
 @Entity()
-@Index(['guildId'])
 export class AnnouncementConfig {
   @PrimaryGeneratedColumn()
   id: number;